Flytt all kode som har med opphørsresultat ut i egen fil (også for tester)
