+++ src/test/enhetstester/kotlin/no/nav/familie/ba/sak/kjerne/behandlingsresultat/BehandlingsresultatUtilsTest.kt
@@ -179,180 +179,6 @@ class BehandlingsresultatUtilsTest {
         )
     }
 
-    @Test
-    fun `hentOpphørsresultatPåBehandling skal returnere IKKE_OPPHØRT dersom nåværende andeler strekker seg lengre enn dagens dato`() {
-        val barn1Aktør = lagPerson(type = PersonType.BARN).aktør
-        val barn2Aktør = lagPerson(type = PersonType.BARN).aktør
-
-        mockkStatic(YearMonth::class)
-        every { YearMonth.now() } returns YearMonth.of(2022, 4)
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mai22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mai22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val nåværendeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mai22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = aug22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val opphørsresultat = BehandlingsresultatUtils.hentOpphørsresultatPåBehandling(nåværendeAndeler, forrigeAndeler)
-
-        assertEquals(BehandlingsresultatUtils.Opphørsresultat.IKKE_OPPHØRT, opphørsresultat)
-    }
-
-    @Test
-    fun `hentOpphørsresultatPåBehandling skal returnere OPPHØRT dersom nåværende andeler opphører mens forrige andeler ikke opphører til og med dagens dato`() {
-        val barn1Aktør = lagPerson(type = PersonType.BARN).aktør
-        val barn2Aktør = lagPerson(type = PersonType.BARN).aktør
-
-        mockkStatic(YearMonth::class)
-        every { YearMonth.now() } returns YearMonth.of(2022, 4)
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = aug22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = aug22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val nåværendeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = feb22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = feb22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val opphørsresultat = BehandlingsresultatUtils.hentOpphørsresultatPåBehandling(nåværendeAndeler, forrigeAndeler)
-
-        assertEquals(BehandlingsresultatUtils.Opphørsresultat.OPPHØRT, opphørsresultat)
-    }
-
-    @Test
-    fun `hentOpphørsresultatPåBehandling skal returnere OPPHØRT dersom nåværende andeler opphører tidligere enn forrige andeler og dagens dato`() {
-        val barn1Aktør = lagPerson(type = PersonType.BARN).aktør
-        val barn2Aktør = lagPerson(type = PersonType.BARN).aktør
-        val apr22 = YearMonth.of(2022, 4)
-
-        mockkStatic(YearMonth::class)
-        every { YearMonth.now() } returns apr22
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mar22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mar22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val nåværendeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = feb22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = feb22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val opphørsresultat = BehandlingsresultatUtils.hentOpphørsresultatPåBehandling(nåværendeAndeler, forrigeAndeler)
-
-        assertEquals(BehandlingsresultatUtils.Opphørsresultat.OPPHØRT, opphørsresultat)
-    }
-
-    @Test
-    fun `hentOpphørsresultatPåBehandling skal returnere FORTSATT_OPPHØRT dersom nåværende andeler har lik opphørsdato som forrige andeler`() {
-        val barn1Aktør = lagPerson(type = PersonType.BARN).aktør
-        val barn2Aktør = lagPerson(type = PersonType.BARN).aktør
-        val apr22 = YearMonth.of(2022, 4)
-
-        mockkStatic(YearMonth::class)
-        every { YearMonth.now() } returns apr22
-
-        val forrigeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mar22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mar22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val nåværendeAndeler = listOf(
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mar22,
-                beløp = 1054,
-                aktør = barn1Aktør
-            ),
-            lagAndelTilkjentYtelse(
-                fom = jan22,
-                tom = mar22,
-                beløp = 1054,
-                aktør = barn2Aktør
-            )
-        )
-
-        val opphørsresultat = BehandlingsresultatUtils.hentOpphørsresultatPåBehandling(nåværendeAndeler, forrigeAndeler)
-
-        assertEquals(BehandlingsresultatUtils.Opphørsresultat.FORTSATT_OPPHØRT, opphørsresultat)
-    }
-
     private fun lagYtelsePerson(
         resultat: YtelsePersonResultat,
         ytelseSlutt: YearMonth? = YearMonth.now().minusMonths(2)
